chore: moving note to under noteFeed
