+++ src/components/Project/NoteFeed/Note/DeleteNote.jsx
@@ -8,7 +8,7 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
 
-import { DeleteNoteMutation as deleteNote } from '../Queries/requests';
+import { DeleteNoteMutation as deleteNote } from '../../Queries/requests';
 
 import { deleteButton } from './Notes.module.scss';
