+++ src/components/Project/NoteFeed/Note/Notes.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/variables.scss';
+@import '../../../../styles/variables.scss';
 
 section {
   display: flex;
